+++ src/l2a_item.cpp
@@ -104,9 +104,9 @@ L2A::Item::Item(const AIArtHandle& placed_item_handle)
 
         if (method_placed_item != method_l2a || alignment_placed_item != alignment_l2a || clip_placed_item != clip_l2a)
         {
-            if (alignment_placed_item == alignment_l2a && property_.GetVersion() < L2A::UTIL::ParseVersion("0.1.0"))
+            if (alignment_placed_item == alignment_l2a && property_.GetVersion() < L2A::UTIL::ParseVersion("1.0.0"))
             {
-                // Starting from v0.1.0 we only support items that are stretched to the boundary box for simplicity
+                // Starting from v1.0.0 we only support items that are stretched to the boundary box for simplicity
                 // reasons. No warning is needed here.
             }
             else
@@ -601,7 +601,7 @@ bool L2A::RedoLaTeXItems(std::vector<L2A::Item>& l2a_items)
     }
 
     // Create the PDFs for the items and store them in the placed items. We dont reset the boundary box here. This is
-    // done in the redo function, we leave it out here, since one might want to use this funciton without resetting the
+    // done in the redo function, we leave it out here, since one might want to use this function without resetting the
     // bounding box.
     for (unsigned int i = 0; i < l2a_items.size(); i++)
     {