+++ src/utils/l2a_execute.cpp
@@ -41,18 +41,10 @@
 /**
  *
  */
-L2A::UTIL::CommandResult L2A::UTIL::ExecuteCommandLine(
-    const ai::UnicodeString& command, const bool quiet, const unsigned long max_time_ms)
+L2A::UTIL::CommandResult L2A::UTIL::ExecuteCommandLine(const ai::UnicodeString& command)
 {
 #ifdef WIN_ENV
-    if (quiet)
-    {
-        return INTERNAL::ExecuteCommandLineWindowsNoConsole(command, max_time_ms);
-    }
-    else
-    {
-        return INTERNAL::ExecuteCommandLineStd(command);
-    }
+    return INTERNAL::ExecuteCommandLineWindowsNoConsole(command);
 #else
     return INTERNAL::ExecuteCommandLineStd(command);
 #endif
@@ -69,6 +61,8 @@ L2A::UTIL::CommandResult L2A::UTIL::INTERNAL::ExecuteCommandLineStd(const ai::Un
 #define popen _popen
 #define pclose _pclose
 #define WEXITSTATUS
+    l2a_error(
+        "ExecuteCommandLineStd is not tested for Windows. If this is adaped, check that unicode works as expected!");
 #endif
     FILE* pipe = popen(L2A::UTIL::StringAiToStd(command).c_str(), "r");
     if (pipe == nullptr)
@@ -98,10 +92,9 @@ L2A::UTIL::CommandResult L2A::UTIL::INTERNAL::ExecuteCommandLineStd(const ai::Un
 /**
  *
  */
-L2A::UTIL::CommandResult L2A::UTIL::INTERNAL::ExecuteCommandLineWindowsNoConsole(
-    const ai::UnicodeString& command, const unsigned long max_time_ms)
+L2A::UTIL::CommandResult L2A::UTIL::INTERNAL::ExecuteCommandLineWindowsNoConsole(const ai::UnicodeString& command)
 {
-#ifndef _WIN32
+#ifndef WIN_ENV
     l2a_error("You are using the function for the wrong OS! Use the system calls via ExecuteCommandLine!");
 #else
     // This code is mainly a combination of
@@ -109,7 +102,8 @@ L2A::UTIL::CommandResult L2A::UTIL::INTERNAL::ExecuteCommandLineWindowsNoConsole
     // https://docs.microsoft.com/en-us/windows/win32/procthread/creating-a-child-process-with-redirected-input-and-output
 
     // Convert the string to platform text.
-    std::string cmdLine = L2A::UTIL::StringAiToStd(command);
+    auto command_wstr = L2A::UTIL::StringAiToStdW(command);
+
     SECURITY_ATTRIBUTES saAttr;
     saAttr.nLength = sizeof(SECURITY_ATTRIBUTES);
     saAttr.bInheritHandle = TRUE;
@@ -127,54 +121,41 @@ L2A::UTIL::CommandResult L2A::UTIL::INTERNAL::ExecuteCommandLineWindowsNoConsole
 
     // Create the process.
     PROCESS_INFORMATION processInformation = {0};
-    STARTUPINFO startupInfo = {0};
+    STARTUPINFOW startupInfo = {0};
     startupInfo.cb = sizeof(startupInfo);
     startupInfo.hStdError = g_hChildStd_OUT_Wr;
     startupInfo.hStdOutput = g_hChildStd_OUT_Wr;
     startupInfo.dwFlags |= STARTF_USESTDHANDLES;
-    BOOL result = CreateProcess(nullptr, (char*)(cmdLine.c_str()), nullptr, nullptr, TRUE,
+    BOOL result = CreateProcessW(nullptr, &command_wstr[0], nullptr, nullptr, TRUE,
         NORMAL_PRIORITY_CLASS | CREATE_NO_WINDOW, nullptr, nullptr, &startupInfo, &processInformation);
 
     // Check if the process could be created.
     if (!result)
     {
         // Get the error from the system
-        LPVOID lpMsgBuf;
-        DWORD dw = GetLastError();
-        FormatMessage(FORMAT_MESSAGE_ALLOCATE_BUFFER | FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS,
-            nullptr, dw, MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT), (LPTSTR)&lpMsgBuf, 0, nullptr);
-
-        // Display the error
-        // CString strError = (LPTSTR) lpMsgBuf;
-        // TRACE(_T("::executeCommandLine() failed at CreateProcess()\nCommand=%s\nMessage=%s\n\n"), cmdLine, strError);
+        LPWSTR lpMsgBuf;
+        const DWORD dw = GetLastError();
+        FormatMessageW(FORMAT_MESSAGE_ALLOCATE_BUFFER | FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS,
+            nullptr, dw, MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT), (LPWSTR)&lpMsgBuf, 0, nullptr);
+        ai::UnicodeString error_string(lpMsgBuf);
 
         // Free resources created by the system
         LocalFree(lpMsgBuf);
 
         // Create error message.
-        l2a_error("Error, process '" + command + "' could not be created!");
+        l2a_error("Error, process '" + command + "' could not be created! Got error: " + error_string);
     }
     else
     {
-        // Successfully created the process.  Wait for it to finish.
-        WaitForSingleObject(processInformation.hProcess, max_time_ms);
-
-        // Get the exit code.
-        DWORD exitCode;
-        result = GetExitCodeProcess(processInformation.hProcess, &exitCode);
-
-        // Close the handles.
-        CloseHandle(processInformation.hProcess);
-        CloseHandle(processInformation.hThread);
+        // Successfully created the process. First close the write handle so the process can finish.
         CloseHandle(g_hChildStd_OUT_Wr);
 
-        // Read the output from the command.
+        // Now we continuously read the ouput of the process
         DWORD dwRead;
         static const int BUFSIZE = 4096;
         CHAR chBuf[BUFSIZE];
         BOOL bSuccess = FALSE;
         std::string result_string = "";
-
         for (;;)
         {
             bSuccess = ReadFile(g_hChildStd_OUT_Rd, chBuf, BUFSIZE, &dwRead, NULL);
@@ -184,12 +165,24 @@ L2A::UTIL::CommandResult L2A::UTIL::INTERNAL::ExecuteCommandLineWindowsNoConsole
             result_string += s;
         }
 
+        // Wait for the process to finish
+        WaitForSingleObject(processInformation.hProcess, INFINITE);
+
+        // Get the exit code
+        DWORD exitCode;
+        result = GetExitCodeProcess(processInformation.hProcess, &exitCode);
+
+        // Close all remaining handles
+        CloseHandle(processInformation.hProcess);
+        CloseHandle(processInformation.hThread);
+        CloseHandle(g_hChildStd_OUT_Rd);
+
         if (!result)
         {
             l2a_error("Executed command but couldn't get exit code.");
         }
 
-        // Return exit code, and command output as unicode string.
+        // Return exit code, and command output as unicode string
         return CommandResult{(int)exitCode, L2A::UTIL::StringStdToAi(result_string)};
     }
 #endif